requirements.txt
Send2Trash==1.8.0
aiohttp==3.7.4
orjson==3.6.4
pydantic==1.8.2
python-dateutil==2.8.2
python-socketio==5.4.1
pytz==2021.3
pyyaml==6.0
requests==2.26.0
